Backend/requirements.txt
flask
flask-pymongo
flask-bcrypt
flask-session
flask-cors
python-dotenv
werkzeug
flask-socketio